feat: clean up terminal screen code
