+++ src/app/terminal/hooks/useTerminalInput.ts
@@ -1,0 +1,13 @@
+import { useState } from "react";
+
+export function useTerminalInput() {
+  const [input, setInput] = useState("");
+
+  const clearInput = () => setInput("");
+
+  return {
+    input,
+    setInput,
+    clearInput,
+  };
+}
