+++ src/app/terminal/index.ts
@@ -1,0 +1,17 @@
+// Components
+export { default as TerminalScreen } from "./components/TerminalScreen";
+export { default as TerminalInput } from "./components/TerminalInput";
+
+// Hooks
+export { useUniverseId } from "./hooks/useUniverseId";
+export { useTerminalHistory } from "./hooks/useTerminalHistory";
+export { useCursor } from "./hooks/useCursor";
+export { useInputFocus } from "./hooks/useInputFocus";
+export { useTerminalInput } from "./hooks/useTerminalInput";
+
+// Services
+export * from "./services/terminalApi";
+
+// Utils
+export * from "./utils/commands";
+export * from "./utils/builtInCommands";
